提交岗位管理代码 解决参数管理、产品管理BUG

+++ itop/src/main/java/com/ei/itop/scmgnt/service/impl/JobServiceImpl.java
@@ -2,16 +2,17 @@ package com.ei.itop.scmgnt.service.impl;
 
 import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 import javax.annotation.Resource;
 
 import org.springframework.stereotype.Service;
 
+import com.ailk.dazzle.exception.BizException;
 import com.ailk.dazzle.util.ibatis.GenericDAO;
+import com.ei.itop.common.dbentity.CcCustProdOp;
 import com.ei.itop.common.dbentity.ScJob;
 import com.ei.itop.common.dbentity.ScJobI18n;
+import com.ei.itop.common.util.SessionUtil;
 import com.ei.itop.scmgnt.bean.JobInfo;
 import com.ei.itop.scmgnt.service.JobService;
 
@@ -27,6 +28,9 @@ public class JobServiceImpl implements JobService{
 	@Resource(name = "app.siCommonDAO")
 	private GenericDAO<Long, JobInfo> jobInfoDAO;
 	
+	@Resource(name = "app.siCommonDAO")
+	private GenericDAO<Long, CcCustProdOp> custProdOpDAO;
+	
 	public List<ScJob> queryJobs(long orgId,String locale) throws Exception {
 		HashMap<String, Object> hm = new HashMap<String, Object>();
 		hm.put("orgId", orgId);
@@ -39,13 +43,7 @@ public class JobServiceImpl implements JobService{
 		HashMap<String, Object> hm = new HashMap<String, Object>();
 		hm.put("orgId", orgId);
 		hm.put("locale", locale);
-		List<JobInfo> jobInfos = jobInfoDAO.findByParams("SC_JOB.selectInfoByOrgId", hm);
-		for(JobInfo ji:jobInfos){
-			hm.clear();
-			hm.put("jobId", ji.getScJobId());
-			ji.setI18n(jobI18nDAO.findByParams("SC_JOB_I18N.selectJobI18n", hm));
-		}
-		return jobInfos;
+		return jobInfoDAO.findByParams("SC_JOB.selectInfoByOrgId", hm);
 	}
 
 	public ScJob getJob(long jobId,String locale) throws Exception {
@@ -74,6 +72,17 @@ public class JobServiceImpl implements JobService{
 	}
 
 	public void removeJob(long jobId) throws Exception {
+		HashMap<String, Object> hm = new HashMap<String, Object>();
+		hm.put("jobId", jobId);
+		List<CcCustProdOp> result = custProdOpDAO.findByParams("CC_CUST_PROD_OP.queryCustProdOpByJobId", hm);
+		if(result!=null&&result.size()>0){
+			throw new BizException(SessionUtil.getRequestContext().getMessage("i18n.scmgnt.jobinfo.JobInUseError"));
+		}
+		//删除国际化数据
+		hm.clear();
+		hm.put("jobId", jobId);
+		jobI18nDAO.deleteByParams("SC_JOB_I18N.deleteByJobId",hm);
+		//删除主数据
 		jobDAO.delete("SC_JOB.deleteByPrimaryKey", jobId);
 	}
 }
